fix(explore): 儲存 503 — 套用 mapNominatimCategory 映射到 pois.type whitelist (v2.14.5)
PR-T (QA round 8) — User: 搜尋的儲存會出現錯誤。Claude 用 /browse 對 prod
測試找到根因。

POST /api/pois/find-or-create 503 → DB CHECK constraint 失敗（pois.type 必
須在 whitelist 內，但 ExplorePage 送 'tourism' / 'amenity' raw category）。

修：抽 mapNominatimCategory 到 src/lib/poiCategory.ts，ExplorePage +
InlineAddPoi 共用。

Verify: tsc clean / 122 files / 1026 tests pass.
